+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to the CDC Aggregator OAI-PMH Repo Handler will be documente
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## TBD - unreleased
+
+### Changed
+
+- Require Kuha OAI-PMH Repo Handler 1.0.2 in requirements.txt.
+- Add primary lookup to `oai_datacite` Publisher from
+  Study.distributors. The current lookup from Study.publishers will
+  remain as a secondary source. (Fixes [#31](https://bitbucket.org/cessda/cessda.cdc.aggregator.oai-pmh-repo-handler/issues/31))
+
+### Fixed
+
+- Change lookup order of preferred PublicationYear value for
+  `oai_datacite`. (Fixes [#30](https://bitbucket.org/cessda/cessda.cdc.aggregator.oai-pmh-repo-handler/issues/30))
+- Format PublicationYear value for `oai_datecite` so that it is a
+  year, instead of full datestamp. (Fixes [#30](https://bitbucket.org/cessda/cessda.cdc.aggregator.oai-pmh-repo-handler/issues/30))
+- Include mandatory Date property to `oai_datacite`. (Fixes [#29](https://bitbucket.org/cessda/cessda.cdc.aggregator.oai-pmh-repo-handler/issues/29))
+
+
+
 ## 0.2.1 - 2022-06-29
 
 ### Changed